+++ src/main/kotlin/com/github/xeonkryptos/integration/gitlab/ui/clone/GitlabLoginData.kt
@@ -1,4 +1,4 @@
-package com.github.xeonkryptos.integration.gitlab.ui.cloneDialog
+package com.github.xeonkryptos.integration.gitlab.ui.clone
 
 data class GitlabLoginData(
     val gitlabHost: String, val gitlabAccessToken: String, val disableCertificateValidation: Boolean